♻️ GH-244 Seal audit-skills permission analysis boundary (I1)
audit/analyze.py imported 12 symbols up from
skills.audit.analyze_permissions, inverting the context boundary
(audit -> skills). Finding I1 of the M5 audit; the other findings
and ADR-0007/0008 already shipped via 560e7063.

Solution:
- Move parsers, classifiers, hygiene auditor, and report writer
  into audit/permissions_model.py (audit context owns the logic)
- Make skills/audit/analyze_permissions.py a thin CLI adapter
  that re-exports the model and keeps the standalone main()
- Repoint audit/analyze.py to import from permissions_model
- Apply effective_cwd() fallback in detect_known_friction
- Move logic tests to tests/audit/test_permissions_model.py;
  add adapter coverage in tests/skills/audit

Fixes: GH-244
