+++ src/dev10x/audit/analyze.py
@@ -7,9 +7,10 @@ shim so the skill-audit pipeline can still run analyze-permissions
 from a Bash entry point.
 
 The implementation reuses the parsers, classifiers, and writers
-already defined in `dev10x.skills.audit.analyze_permissions` —
-this factory is the seam that lets MCP callers skip the subprocess
-hop and consume the report as data.
+defined in `dev10x.audit.permissions_model` — keeping the analysis
+logic inside the audit context (GH-244, I1 / ADR-0008) instead of
+importing up into the skills layer. This factory is the seam that lets
+MCP callers skip the subprocess hop and consume the report as data.
 """
 
 from __future__ import annotations
@@ -18,8 +19,7 @@ import io
 from dataclasses import dataclass, field
 from pathlib import Path
 
-from dev10x.domain.claude_paths import ClaudeDir
-from dev10x.skills.audit.analyze_permissions import (
+from dev10x.audit.permissions_model import (
     Finding,
     HygieneFinding,
     audit_script_hygiene,
@@ -31,9 +31,10 @@ from dev10x.skills.audit.analyze_permissions import (
     propose_allow_rules,
     write_output,
 )
-from dev10x.skills.audit.analyze_permissions import (
+from dev10x.audit.permissions_model import (
     analyze_permissions as _analyze_permissions,
 )
+from dev10x.domain.claude_paths import ClaudeDir
 from dev10x.subprocess_utils import effective_cwd
 
 
@@ -78,11 +79,10 @@ def build_audit_report(
     findings = _analyze_permissions(calls=calls, rules=rules)
     findings = count_nuisance_patterns(findings=findings)
 
-    # GH-979 (H6): analyze_permissions is a deps-less standalone uv-script
-    # that cannot import dev10x, so the effective-CWD fallback lands here at
-    # the importing seam. When an MCP caller bound the worktree via use_cwd,
-    # default the project root to it; standalone CLI callers pass None and
-    # detect_known_friction falls back to os.getcwd() as before.
+    # GH-979 (H6): when an MCP caller bound the worktree via use_cwd, default
+    # the project root to it at this seam. permissions_model.detect_known_friction
+    # applies the same effective_cwd() fallback for standalone CLI callers that
+    # pass None.
     extra = detect_known_friction(
         calls=calls,
         additional_dirs=additional_dirs,
